+++ src/css/mod.rs
@@ -1,5 +1,6 @@
 //! Types for css values and rules.
 mod atrule;
+mod binop;
 mod call_args;
 mod comment;
 mod item;
@@ -11,12 +12,13 @@ mod value;
 mod valueformat;
 
 pub use self::atrule::{AtRule, AtRuleBodyItem};
+pub use self::binop::BinOp;
 pub use self::call_args::CallArgs;
 pub use self::comment::Comment;
 pub use self::item::{Import, Item};
 pub use self::rule::{BodyItem, Property, Rule};
 pub use self::selectors::{BadSelector, Selector, SelectorPart, Selectors};
 pub use self::string::CssString;
-pub use self::value::{BinOp, Value, ValueMap, ValueToMapError};
+pub use self::value::{InvalidCss, Value, ValueMap, ValueToMapError};
 
 pub(crate) use self::util::{is_calc_name, is_function_name, is_not};