+++ src/value/mod.rs
@@ -13,7 +13,7 @@ pub use self::colors::{Color, Hsla, Hwba, RgbFormat, Rgba};
 pub use self::list_separator::ListSeparator;
 pub use self::number::{BadNumber, Number, Rational};
 pub use self::numeric::Numeric;
-pub use self::operator::Operator;
+pub use self::operator::{BadOp, Operator};
 pub use self::quotes::Quotes;
 pub use self::unit::{CssDimension, Dimension, Unit};
 pub use self::unitset::UnitSet;