+++ tests/spec/libsass_closed_issues/issue_2147/rhs.rs
@@ -6,7 +6,6 @@ fn runner() -> crate::TestRunner {
 }
 
 #[test]
-#[ignore] // missing error
 fn test() {
     assert_eq!(
         runner().err("$map: 1 + (a:b,c:d);\r\n"),