+++ tests/spec/non_conformant/sass_4_0/color_arithmetic/division/color_number.rs
@@ -6,7 +6,6 @@ fn runner() -> crate::TestRunner {
 }
 
 #[test]
-#[ignore] // missing error
 fn test() {
     assert_eq!(
         runner().err("$v: #abc / 1;\n"),