+++ tests/spec/non_conformant/sass_4_0/color_arithmetic/subtraction/color_color.rs
@@ -6,7 +6,6 @@ fn runner() -> crate::TestRunner {
 }
 
 #[test]
-#[ignore] // missing error
 fn test() {
     assert_eq!(
         runner().err("$v: #abc - #123;\n"),